refactor: reorganiza medição de tempo e busca hash
