+++ estrutura-dados-busca-sequencial/src/tempo.c
@@ -1,9 +1,11 @@
 #include <time.h>
 #include "tempo.h"
 #include "busca.h"
+#include "hash.h"
 
-//essa função calcula quanto tempo leva uma busca em média
-double calcular_tempo_total(Produto* vetor, int tamanho, int*ids_busca, int qtd_buscas) {
+//essa função calcula o tempo total de uma busca sequencial
+double calcular_tempo_sequencial(Produto* vetor, int tamanho, int*ids_busca, int qtd_buscas) 
+{
     clock_t inicio, fim;
 
     //momento em que o teste começa (conta em ticks de CPU)
@@ -15,5 +17,19 @@ double calcular_tempo_total(Produto* vetor, int tamanho, int*ids_busca, int qtd_
     fim = clock();
 
     //calcula a diferença entre fim e início, CLOCKS_PER_SEC converte os ticks de CPU em segundos
+    return (double)(fim - inicio) / CLOCKS_PER_SEC;
+}
+
+//essa função calcula o tempo total de uma busca na tabela hash
+double calcular_tempo_hash(TabelaHash *hash, int *ids_busca, int qtd_buscas)
+{
+    clock_t inicio, fim;
+    inicio = clock();
+
+    for (int i = 0; i < qtd_buscas; i++) {  //executa a busca mesmo quando o ID não existe
+        buscar_hash(hash, ids_busca[i]);    //retorna NULL para ID < 0 
+    }
+
+    fim = clock();
     return (double)(fim - inicio) / CLOCKS_PER_SEC;
 }
