--- estrutura-dados-busca-sequencial/src/tempo_hash.h
@@ -1,7 +1,0 @@
-#ifndef TEMPO_HASH_H
-#define TEMPO_HASH_H
-#include "hash.h"
-
-double calcular_tempo_hash(TabelaHash *hash, int *ids_busca, int qtd_buscas);
-
-#endif
